requirements.txt

pyodbc==5.1.0
python-dotenv==1.0.1
httpx==0.27.2
tenacity==9.0.0
loguru==0.7.2
pydantic==2.8.2
